Refactor to seperate-out dao classes from the store classes
