+++ components/email-mgt/org.wso2.carbon.email.mgt/src/test/resources/testng.xml
@@ -20,7 +20,7 @@
 
     <test name="EmailMgtTests" preserve-order="true" parallel="false">
         <classes>
-            <class name="org.wso2.carbon.email.mgt.dao.TemplatePersistenceManagerFactoryTest"/>
+            <class name="org.wso2.carbon.email.mgt.store.TemplatePersistenceManagerFactoryTest"/>
             <class name="org.wso2.carbon.email.mgt.OrganizationEmailTemplateTest"/>
             <class name="org.wso2.carbon.email.mgt.ApplicationEmailTemplateTest"/>
             <class name="org.wso2.carbon.email.mgt.util.I18nEmailUtilTest"/>
